Lf network refactoring
