+++ simple-loadflow/src/main/java/com/powsybl/loadflow/simple/network/impl/LfBusImpl.java
@@ -4,11 +4,14 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.loadflow.simple.network;
+package com.powsybl.loadflow.simple.network.impl;
 
 import com.powsybl.commons.PowsyblException;
 import com.powsybl.iidm.network.*;
 import com.powsybl.iidm.network.extensions.ActivePowerControl;
+import com.powsybl.loadflow.simple.network.AbstractLfBus;
+import com.powsybl.loadflow.simple.network.LfShunt;
+import com.powsybl.loadflow.simple.network.PerUnit;
 
 import java.util.*;
