+++ simple-loadflow/src/main/java/com/powsybl/loadflow/simple/network/impl/LfLeg2or3Branch.java
@@ -4,9 +4,13 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.loadflow.simple.network;
+package com.powsybl.loadflow.simple.network.impl;
 
 import com.powsybl.iidm.network.ThreeWindingsTransformer;
+import com.powsybl.loadflow.simple.network.AbstractLfBranch;
+import com.powsybl.loadflow.simple.network.LfBus;
+import com.powsybl.loadflow.simple.network.PerUnit;
+import com.powsybl.loadflow.simple.network.PiModel;
 
 import java.util.Objects;
