+++ simple-loadflow/src/main/java/com/powsybl/loadflow/simple/network/impl/LfNetworks.java
@@ -4,22 +4,18 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.loadflow.simple.network;
+package com.powsybl.loadflow.simple.network.impl;
 
-import com.fasterxml.jackson.core.JsonFactory;
-import com.fasterxml.jackson.core.JsonGenerator;
 import com.google.common.base.Stopwatch;
 import com.powsybl.commons.PowsyblException;
 import com.powsybl.iidm.network.*;
+import com.powsybl.loadflow.simple.network.LfBranch;
+import com.powsybl.loadflow.simple.network.LfBus;
+import com.powsybl.loadflow.simple.network.LfNetwork;
+import com.powsybl.loadflow.simple.network.SlackBusSelector;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.io.IOException;
-import java.io.UncheckedIOException;
-import java.io.Writer;
-import java.nio.charset.StandardCharsets;
-import java.nio.file.Files;
-import java.nio.file.Path;
 import java.util.*;
 import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
@@ -27,17 +23,9 @@ import java.util.stream.Collectors;
 /**
  * @author Geoffroy Jamgotchian <geoffroy.jamgotchian at rte-france.com>
  */
-public class NetworkContext {
+public final class LfNetworks {
 
-    private static final Logger LOGGER = LoggerFactory.getLogger(NetworkContext.class);
-
-    private final Network network;
-
-    private final List<LfBus> buses;
-
-    private final List<LfBranch> branches;
-
-    private final LfBus slackBus;
+    private static final Logger LOGGER = LoggerFactory.getLogger(LfNetworks.class);
 
     private static class CreationContext {
 
@@ -50,22 +38,7 @@ public class NetworkContext {
         private final Map<String, Integer> busIdToNum = new HashMap<>();
     }
 
-    public NetworkContext(Network network, List<Bus> buses, SlackBusSelector slackBusSelector, Map<HvdcConverterStation, HvdcLine> hvdcLines) {
-        this.network = Objects.requireNonNull(network);
-        Objects.requireNonNull(buses);
-        if (buses.isEmpty()) {
-            throw new IllegalArgumentException("Empty bus list");
-        }
-        Objects.requireNonNull(slackBusSelector);
-        Objects.requireNonNull(hvdcLines);
-
-        CreationContext creationContext = new CreationContext();
-        this.buses = createBuses(buses, hvdcLines, creationContext);
-        branches = createBranches(this.buses, creationContext);
-
-        slackBus = slackBusSelector.select(this.buses);
-        slackBus.setSlack(true);
-        LOGGER.debug("Selected slack bus (class={}): {}", slackBusSelector.getClass().getSimpleName(), slackBus.getId());
+    private LfNetworks() {
     }
 
     private static List<LfBus> createBuses(List<Bus> buses, Map<HvdcConverterStation, HvdcLine> hvdcLines, CreationContext creationContext) {
@@ -221,7 +194,15 @@ public class NetworkContext {
         return lfBus;
     }
 
-    public static List<NetworkContext> of(Network network, SlackBusSelector slackBusSelector) {
+    private static LfNetwork create(List<Bus> buses, SlackBusSelector slackBusSelector,
+                                    Map<HvdcConverterStation, HvdcLine> hvdcLines) {
+        CreationContext creationContext = new CreationContext();
+        List<LfBus> lfBuses = createBuses(buses, hvdcLines, creationContext);
+        List<LfBranch> lfBranches = createBranches(lfBuses, creationContext);
+        return new LfNetwork(lfBuses, lfBranches, slackBusSelector);
+    }
+
+    public static List<LfNetwork> create(Network network, SlackBusSelector slackBusSelector) {
         Objects.requireNonNull(network);
         Map<Integer, List<Bus>> buseByCc = new TreeMap<>();
         for (Bus bus : network.getBusView().getBuses()) {
@@ -240,7 +221,7 @@ public class NetworkContext {
 
         return buseByCc.entrySet().stream()
                 .filter(e -> e.getKey() == ComponentConstants.MAIN_NUM)
-                .map(e -> new NetworkContext(network, e.getValue(), slackBusSelector, hvdcLines))
+                .map(e -> create(e.getValue(), slackBusSelector, hvdcLines))
                 .collect(Collectors.toList());
     }
 
@@ -265,163 +246,4 @@ public class NetworkContext {
         LOGGER.debug("Network reset done in {} ms", stopwatch.elapsed(TimeUnit.MILLISECONDS));
     }
 
-    public List<LfBranch> getBranches() {
-        return branches;
-    }
-
-    public List<LfBus> getBuses() {
-        return buses;
-    }
-
-    public LfBus getBus(int num) {
-        return buses.get(num);
-    }
-
-    public LfBus getSlackBus() {
-        return slackBus;
-    }
-
-    public void updateState() {
-        for (LfBus bus : buses) {
-            bus.updateState();
-            for (LfShunt shunt : bus.getShunts()) {
-                shunt.updateState();
-            }
-        }
-        for (LfBranch branch : branches) {
-            branch.updateState();
-        }
-    }
-
-    public void writeJson(Path file) {
-        try (Writer writer = Files.newBufferedWriter(file, StandardCharsets.UTF_8)) {
-            writeJson(writer);
-        } catch (IOException e) {
-            throw new UncheckedIOException(e);
-        }
-    }
-
-    public void writeJson(LfBus bus, JsonGenerator jsonGenerator) throws IOException {
-        jsonGenerator.writeStringField("id", bus.getId());
-        jsonGenerator.writeNumberField("num", bus.getNum());
-        if (bus.getGenerationTargetP() != 0) {
-            jsonGenerator.writeNumberField("generationTargetP", bus.getGenerationTargetP());
-        }
-        if (bus.getGenerationTargetQ() != 0) {
-            jsonGenerator.writeNumberField("generationTargetQ", bus.getGenerationTargetQ());
-        }
-        if (bus.getLoadTargetP() != 0) {
-            jsonGenerator.writeNumberField("loadTargetP", bus.getLoadTargetP());
-        }
-        if (bus.getLoadTargetQ() != 0) {
-            jsonGenerator.writeNumberField("loadTargetQ", bus.getLoadTargetQ());
-        }
-        if (!Double.isNaN(bus.getMinP())) {
-            jsonGenerator.writeNumberField("minP", bus.getMinP());
-        }
-        if (!Double.isNaN(bus.getMaxP())) {
-            jsonGenerator.writeNumberField("maxP", bus.getMaxP());
-        }
-        if (!Double.isNaN(bus.getTargetV())) {
-            jsonGenerator.writeNumberField("targetV", bus.getTargetV());
-        }
-        if (!Double.isNaN(bus.getV())) {
-            jsonGenerator.writeNumberField("v", bus.getV());
-        }
-        if (!Double.isNaN(bus.getAngle())) {
-            jsonGenerator.writeNumberField("angle", bus.getAngle());
-        }
-    }
-
-    public void writeJson(LfBranch branch, JsonGenerator jsonGenerator) throws IOException {
-        LfBus bus1 = branch.getBus1();
-        LfBus bus2 = branch.getBus2();
-        if (bus1 != null) {
-            jsonGenerator.writeNumberField("num1", bus1.getNum());
-        }
-        if (bus2 != null) {
-            jsonGenerator.writeNumberField("num2", bus2.getNum());
-        }
-        jsonGenerator.writeNumberField("x", branch.x());
-        jsonGenerator.writeNumberField("y", branch.y());
-        jsonGenerator.writeNumberField("ksi", branch.ksi());
-        if (branch.g1() != 0) {
-            jsonGenerator.writeNumberField("g1", branch.g1());
-        }
-        if (branch.g2() != 0) {
-            jsonGenerator.writeNumberField("g2", branch.g2());
-        }
-        if (branch.b1() != 0) {
-            jsonGenerator.writeNumberField("b1", branch.b1());
-        }
-        if (branch.b2() != 0) {
-            jsonGenerator.writeNumberField("b2", branch.b2());
-        }
-        if (branch.r1() != 1) {
-            jsonGenerator.writeNumberField("r1", branch.r1());
-        }
-        if (branch.r2() != 1) {
-            jsonGenerator.writeNumberField("r2", branch.r2());
-        }
-        if (branch.a1() != 0) {
-            jsonGenerator.writeNumberField("a1", branch.a1());
-        }
-        if (branch.a2() != 0) {
-            jsonGenerator.writeNumberField("a2", branch.a2());
-        }
-    }
-
-    public void writeJson(LfShunt shunt, JsonGenerator jsonGenerator) throws IOException {
-        jsonGenerator.writeNumberField("b", shunt.getB());
-    }
-
-    public void writeJson(Writer writer) {
-        Objects.requireNonNull(writer);
-        try (JsonGenerator jsonGenerator = new JsonFactory()
-                .createGenerator(writer)
-                .useDefaultPrettyPrinter()) {
-            jsonGenerator.writeStartObject();
-
-            jsonGenerator.writeFieldName("buses");
-            jsonGenerator.writeStartArray();
-            for (LfBus bus : buses) {
-                jsonGenerator.writeStartObject();
-
-                writeJson(bus, jsonGenerator);
-
-                List<LfShunt> shunts = bus.getShunts();
-                if (!shunts.isEmpty()) {
-                    jsonGenerator.writeFieldName("shunts");
-                    jsonGenerator.writeStartArray();
-                    for (LfShunt shunt : shunts) {
-                        jsonGenerator.writeStartObject();
-
-                        writeJson(shunt, jsonGenerator);
-
-                        jsonGenerator.writeEndObject();
-                    }
-                    jsonGenerator.writeEndArray();
-                }
-
-                jsonGenerator.writeEndObject();
-            }
-            jsonGenerator.writeEndArray();
-
-            jsonGenerator.writeFieldName("branches");
-            jsonGenerator.writeStartArray();
-            for (LfBranch branch : branches) {
-                jsonGenerator.writeStartObject();
-
-                writeJson(branch, jsonGenerator);
-
-                jsonGenerator.writeEndObject();
-            }
-            jsonGenerator.writeEndArray();
-
-            jsonGenerator.writeEndObject();
-        } catch (IOException e) {
-            throw new UncheckedIOException(e);
-        }
-    }
 }
-
